backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
Flask-JWT-Extended==4.7.1
Flask-Cors==5.0.1
Flask-Migrate==4.1.0
python-dotenv==1.1.0
Werkzeug==3.1.3
marshmallow==3.25.1
